backend/requirements.txt
fastapi==0.107.0
uvicorn[standard]==0.23.2
requests==2.31.0
beautifulsoup4==4.12.2
transformers==4.42.0
sentence-transformers==2.2.2
torch==2.1.0
weaviate-client==3.22.0
python-dotenv==1.0.0
nltk==3.8.1
python-multipart==0.0.6
pydantic>=2.5.0,<3.0.0
typing_extensions==4.8.0
tqdm==4.66.1